Refactored variant into its own repo
