+++ agr_api/src/main/java/org/alliancegenome/api/application/RestDefaultObjectMapper.java
@@ -1,12 +1,13 @@
 package org.alliancegenome.api.application;
 
+import javax.ws.rs.ext.ContextResolver;
+import javax.ws.rs.ext.Provider;
+
+import org.alliancegenome.core.config.ConfigHelper;
+
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import org.alliancegenome.core.config.ConfigHelper;
-
-import javax.ws.rs.ext.ContextResolver;
-import javax.ws.rs.ext.Provider;
 
 @Provider
 public class RestDefaultObjectMapper implements ContextResolver<ObjectMapper> {
