+++ agr_api/src/main/java/org/alliancegenome/api/service/VariantService.java
@@ -1,25 +1,23 @@
 package org.alliancegenome.api.service;
 
-import org.alliancegenome.cache.repository.helper.JsonResultResponse;
-import org.alliancegenome.es.model.query.Pagination;
-import org.alliancegenome.neo4j.entity.node.Allele;
-import org.alliancegenome.neo4j.entity.node.Transcript;
-import org.alliancegenome.neo4j.entity.node.Variant;
-import org.alliancegenome.neo4j.repository.AlleleRepository;
+import java.time.LocalDateTime;
+import java.util.List;
 
 import javax.enterprise.context.RequestScoped;
+
+import org.alliancegenome.cache.repository.helper.JsonResultResponse;
+import org.alliancegenome.es.model.query.Pagination;
+import org.alliancegenome.neo4j.entity.node.Transcript;
+import org.alliancegenome.neo4j.entity.node.Variant;
+import org.alliancegenome.neo4j.repository.VariantRepository;
 
 @RequestScoped
 public class VariantService {
 
-    private AlleleRepository alleleRepo = new AlleleRepository();
-
-    public Allele getById(String id) {
-        return alleleRepo.getAllele(id);
-    }
+    private VariantRepository variantRepo = new VariantRepository();
 
     public JsonResultResponse<Transcript> getTranscriptsByVariant(String variantID, Pagination pagination) {
-        Variant variant = alleleRepo.getVariant(variantID);
+        Variant variant = variantRepo.getVariant(variantID);
 
         JsonResultResponse<Transcript> response = new JsonResultResponse<>();
         if (variant != null) {
@@ -29,4 +27,17 @@ public class VariantService {
         return response;
     }
 
+    public JsonResultResponse<Variant> getVariants(String id, Pagination pagination) {
+        LocalDateTime startDate = LocalDateTime.now();
+
+        List<Variant> variants = variantRepo.getVariantsOfAllele(id);
+
+        JsonResultResponse<Variant> result = new JsonResultResponse<>();
+
+        FilterService<Variant> filterService = new FilterService<>(null);
+        result.setTotal(variants.size());
+        result.setResults(filterService.getPaginatedAnnotations(pagination, variants));
+        result.calculateRequestDuration(startDate);
+        return result;
+    }
 }
