+++ agr_java_core/src/main/java/org/alliancegenome/neo4j/repository/AlleleRepository.java
@@ -78,42 +78,4 @@ public class AlleleRepository extends Neo4jRepository<Allele> {
                 .collect(Collectors.toSet());
     }
 
-    public List<Variant> getVariants(String id) {
-        HashMap<String, String> map = new HashMap<>();
-        String paramName = "alleleID";
-        map.put(paramName, id);
-        String query = "";
-        query += " MATCH p1=(a:Allele)<-[:VARIATION]-(variant:Variant)--(soTerm:SOTerm) ";
-        query += " WHERE a.primaryKey = {" + paramName + "}";
-        query += " OPTIONAL MATCH consequence=(:GeneLevelConsequence)<-[:ASSOCIATION]-(variant:Variant)";
-        query += " OPTIONAL MATCH loc=(variant:Variant)-[:ASSOCIATION]->(:GenomicLocation)-[:ASSOCIATION]->(:Chromosome)";
-        query += " OPTIONAL MATCH p2=(variant:Variant)<-[:COMPUTED_GENE]-(:Gene)-[:ASSOCIATION]->(:GenomicLocation)-[:ASSOCIATION]->(:Chromosome)";
-        query += " RETURN p1, p2, loc, consequence ";
-
-        Iterable<Variant> alleles = query(Variant.class, query, map);
-        return StreamSupport.stream(alleles.spliterator(), false)
-                .collect(Collectors.toList());
-
-
-    }
-
-    public Variant getVariant(String variantID) {
-        HashMap<String, String> map = new HashMap<>();
-        String paramName = "variantID";
-        map.put(paramName, variantID);
-        String query = "";
-        query += " MATCH p1=(t:Transcript)-[:ASSOCIATION]->(variant:Variant)--(soTerm:SOTerm) ";
-        query += " WHERE variant.primaryKey = {" + paramName + "}";
-        query += " OPTIONAL MATCH consequence=(:TranscriptLevelConsequence)<-[:ASSOCIATION]-(t:Transcript)<-[:TRANSCRIPT_TYPE]-(:SOTerm)";
-        query += " OPTIONAL MATCH gene=(t:Transcript)-[:TRANSCRIPT]-(:Gene)";
-        query += " RETURN p1, consequence, gene ";
-
-        Iterable<Variant> variants = query(Variant.class, query, map);
-        for (Variant a : variants) {
-            if (a.getPrimaryKey().equals(variantID)) {
-                return a;
-            }
-        }
-        return null;
-    }
 }
